refactor(api-edge): Zonen-Projektion in eine geteilte Util ziehen
`zonedParts` lag zweimal im Edge: vollstaendig in `workers/scheduled-slot.ts`
und, auf den Kalendertag reduziert, in `utils/business-day-date.ts`. Beide
Stellen hatten unabhaengig voneinander dieselbe Entscheidung gegen den
`toLocaleString`-Roundtrip getroffen; eine dritte Kopie stand mit der
Oeffnungszeiten-Pruefung der Vorbestellungen an.

Reiner Move — die Specs beider Dateien laufen unveraendert durch (33 Tests).

Refs #279
